third_party/WebKit/LayoutTests/svg/animations/svgnumber-animation-2-expected.txt

SVG 1.1 dynamic animation tests
Test for SVGNumber animation with different exponents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parseFloat(getComputedStyle(rect).opacity) is 0
PASS parseFloat(getComputedStyle(rect).opacity) is 0.5
PASS parseFloat(getComputedStyle(rect).opacity) is 1
PASS parseFloat(getComputedStyle(rect).opacity) is 0
PASS successfullyParsed is true

TEST COMPLETE
